Refactored fetchOrderHistory
